refactor(dto): reorganize dto packages from domain-based to request/response
