+++ src/main/java/roomescape/controller/dto/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto.reservation;
+package roomescape.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
