+++ src/main/java/roomescape/controller/dto/response/AdminReservationsResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto.reservation;
+package roomescape.controller.dto.response;
 
 import java.util.List;
