+++ src/main/java/roomescape/controller/dto/response/AvailableReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto.time;
+package roomescape.controller.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
