+++ src/main/java/roomescape/controller/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto.reservation;
+package roomescape.controller.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
