+++ src/main/java/roomescape/controller/dto/response/ThemesResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto.theme;
+package roomescape.controller.dto.response;
 
 import java.util.List;
